+++ src/edu/uwo/csd/dcsim2/application/WorkConsumer.java
@@ -2,6 +2,6 @@ package edu.uwo.csd.dcsim2.application;
 
 public interface WorkConsumer {
 
-	public void addWork(int work);
+	public void addWork(double work);
 	
 }
